update experiment structure (#94)
* the stimulus text is now not displayed when it is played

* updated view to use float in the wait function

* fixed a test warning

* moved constants out of experiment into main

* updated diagrams to match

+++ auditory_stimulation/view/view.py
@@ -64,6 +64,6 @@ class AView(AObserver):
         ...
 
     @abstractmethod
-    def wait(self, secs: int) -> None:
+    def wait(self, secs: float) -> None:
         """Wait for the specified amount of time."""
         ...
